+++ packages/react-api/src/types.d.ts
@@ -43,4 +43,4 @@ export type UseCartoLayerFilterProps = {
   maskEnabled: boolean;
 } & SourceProps;
 
-export type ExecuteSQLResponse = Promise<FeatureCollection | {}[]>;
+export type ExecuteSQLResponse<Response = FeatureCollection | {}[]> = Promise<Response>;