+++ packages/react-widgets/src/widgets/CategoryWidget.js
@@ -23,6 +23,7 @@ const EMPTY_ARRAY = [];
  * @param  {Function} [props.formatter] - Function to format each value returned.
  * @param  {Object} [props.labels] - Overwrite category labels.
  * @param  {boolean} [props.animation] - Enable/disable widget animations on data updates. Enabled by default.
+ * @param  {boolean} [props.filterable] - Enable/disable widget filtering capabilities. Enabled by default.
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
  * @param  {Object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]()
@@ -38,6 +39,7 @@ function CategoryWidget(props) {
     formatter,
     labels,
     animation,
+    filterable,
     onError,
     wrapperProps,
     noDataAlertProps
@@ -125,6 +127,7 @@ function CategoryWidget(props) {
           selectedCategories={selectedCategories}
           onSelectedCategoriesChange={handleSelectedCategoriesChange}
           animation={animation}
+          filterable={filterable}
         />
       ) : (
         <NoDataAlert {...noDataAlertProps} />
@@ -143,6 +146,7 @@ CategoryWidget.propTypes = {
   formatter: PropTypes.func,
   labels: PropTypes.object,
   animation: PropTypes.bool,
+  filterable: PropTypes.bool,
   onError: PropTypes.func,
   wrapperProps: PropTypes.object,
   noDataAlertProps: PropTypes.object
@@ -151,6 +155,7 @@ CategoryWidget.propTypes = {
 CategoryWidget.defaultProps = {
   labels: {},
   animation: true,
+  filterable: true,
   wrapperProps: {},
   noDataAlertProps: {}
 };