+++ packages/react-widgets/src/widgets/TableWidget.js
@@ -16,6 +16,9 @@ import { selectAreFeaturesReadyForSource } from '@carto/react-redux/';
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  * @param  {Object} [props.wrapperProps] - Extra props to pass to [WrapperWidgetUI](https://storybook-react.carto.com/?path=/docs/widgets-wrapperwidgetui--default)
  * @param  {Object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]()
+ * @param  {number} [props.initialPageSize] - Initial number of rows per page
+ * @param  {string} [props.height] - Static widget height, required for scrollable table content
+ * @param  {boolean} [props.dense] - Whether the table should use a compact layout with smaller cell paddings
  */
 function TableWidget({
   id,
@@ -24,9 +27,12 @@ function TableWidget({
   columns,
   wrapperProps,
   noDataAlertProps,
-  onError
+  onError,
+  initialPageSize = 10,
+  height,
+  dense
 }) {
-  const [rowsPerPage, setRowsPerPage] = useState(10);
+  const [rowsPerPage, setRowsPerPage] = useState(initialPageSize);
   const [page, setPage] = useState(0);
   const [totalCount, setTotalCount] = useState(0);
 
@@ -98,6 +104,8 @@ function TableWidget({
           sortDirection={sortDirection}
           onSetSortBy={setSortBy}
           onSetSortDirection={setSortDirection}
+          height={height}
+          dense={dense}
         />
       ) : (
         <NoDataAlert {...noDataAlertProps} />
@@ -119,7 +127,10 @@ TableWidget.propTypes = {
   ).isRequired,
   onError: PropTypes.func,
   wrapperProps: PropTypes.object,
-  noDataAlertProps: PropTypes.object
+  noDataAlertProps: PropTypes.object,
+  initialPageSize: PropTypes.number,
+  height: PropTypes.string,
+  dense: PropTypes.bool
 };
 
 export default TableWidget;